handlebars and stuff when working with zach TA
