+++ controllers/api/gameRoutes.js
@@ -3,18 +3,6 @@ const { Games } = require('../../models');
 
 // current path /api/games
 
-// get all games
-router.get("/", async (req, res) => {
-  try {
-    const gameData = await Games.findAll();
-    const games = gameData.map((game) => game.get({plain: true}))
-    res.render("homepage", {games});
-  } catch (err) {
-    console.log(err);
-    res.status(500).json(err);
-  }
-});
-
 // add game
 router.post("/addgame", (req, res) => {
   console.info(`${req.method} request received to add Games.`);
