+++ controllers/homeRoutes.js
@@ -1,9 +1,16 @@
 const router = require('express').Router();
+const { Games } = require('../models');
 
-// main route
+// get all games
 router.get("/", async (req, res) => {
-  res.sendFile(path.join(__dirname, "../views/index.html"));
+  try {
+    const gameData = await Games.findAll();
+    const games = gameData.map((game) => game.get({plain: true}))
+    res.render("homepage", {games});
+  } catch (err) {
+    console.log(err);
+    res.status(500).json(err);
+  }
 });
 
 module.exports = router;
-
