+++ controllers/routes.js
@@ -1,7 +1,10 @@
 const router = require('express').Router();
+const homeRoutes = require('./homeRoutes')
 
 const apiRoutes = require('./api');
 
 router.use('/api', apiRoutes);
+router.use(homeRoutes)
 
 module.exports = router;
+
