+++ views/partials/gamedetails.handlebars
@@ -1,0 +1,24 @@
+                <div class="card m-6" style="width: 18rem;">
+                <header class="card-header has-background-primary">
+                    <button class="card-header-icon" aria-label="more options">
+                        <span class="icon">
+                            <i class="fas fa-dice-d20 fa-xl" aria-hidden="true"></i>
+                        </span>
+                    <p class="card-header-title subtitle has-text-centered" id="#gameTitle">
+                        {{gameName}}
+                    </p>
+                    </button>
+                </header>
+                <div class="card-content" style=" border-style: solid;">
+                    <div class="content">
+                        <h5 class="has-text-centered" id="#gameType">{{gameType}}</h5>
+                        <li id="gamePlayers">Players: 1-4</li>
+                        <li id="gameLength">Game Length: (Short, Med, Long)</li>
+                        <p id="#gameDesc">{{gameDescription}}</p>
+                    </div>
+                </div>
+                <footer class="card-footer">
+                    <a href="#" class="card-footer-item is-outlined">Favorite</a>
+                    <a href="#" class="card-footer-item">Remove</a>
+                </footer>
+            </div>
